edu_cli/requirements.txt
google-generativeai>=0.3.0
rich>=10.0.0
prompt-toolkit>=3.0.0
pyfiglet>=0.8.post1
